requirements.txt
aiogram==2.20
aiogram-dialog==1.4.1
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
Babel==2.9.1
cachetools==4.2.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.5
django-environ==0.8.1
django-rest-auth-bearer==0.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Flask==2.1.2
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.7
MarkupSafe==2.1.1
multidict==6.0.2
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
uWSGI==2.0.20
Werkzeug==2.1.2
yarl==1.7.2
